added info and contact pages + form says thank you
contact page is a bit redundant rn but thats ok.
im also tinkering on a lightmode button and menu stuff but that's not important right now.
i kinda want to overhaul the page layout so it's cleaner when you collapse it on desktop - submenu could just be part of the header and it's a half-half situation. will see once i have the time but for now this is good enough.